+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/search/SearchUsers.kt
@@ -4,11 +4,16 @@ import androidx.compose.foundation.lazy.LazyListScope
 import de.connect2x.trixnity.messenger.util.Search
 import de.connect2x.trixnity.messenger.util.UserSearchHandler
 
-fun LazyListScope.SearchUsersLocally(searchHandler: UserSearchHandler, onUserClick: (Search.SearchUserElement) -> Unit, userSearch: UserSearchResultListView, searchResults: UserSearchResultListView.SearchResultState) {
+fun LazyListScope.searchUsersLocally(
+    searchHandler: UserSearchHandler,
+    onUserClick: (Search.SearchUserElement) -> Unit,
+    searchResults: SearchResultState,
+    userSearchResultListView: UserSearchResultListView,
+) {
     stickyHeader(key = "UserSearchField") {
         UserSearchField(searchHandler)
     }
-    userSearch.createLazyComposables(this, searchResults) {
+    userSearchResultListView.create(this, searchResults) {
         onUserClick(it)
     }
 }